Fix warnings up to c++14
Notes:
* the implementation of strftime deemed broken is the one actually
  conforming to the standard
* the --verbose flag of gsmsiexfer actually wasn't used, so it was
  ifdef-ed out
* _debug / env=GSM_DEBUG in UnixSerialPort seems to be superseded by
  GSMLIB_DEBUG env
* the timerMtx mutex in UnixSerialPort was aready unused because of the
  defines; see the comment in code for details

+++ gsmlib/gsm_me_ta.cc
@@ -19,6 +19,7 @@
 #include <gsmlib/gsm_parser.h>
 #include <gsmlib/gsm_sysdep.h>
 
+#include <cstdint>
 #include <cstdlib>
 
 using namespace gsmlib;
@@ -1019,8 +1020,8 @@ void MeTa::sendSMSs(Ref<SMSSubmitMessage> smsTemplate, std::string text,
     {
       if (concatenatedMessageId != -1)
       {
-        unsigned char udhs[] = {0x00, 0x03, concatenatedMessageId,
-                                numMessages, ++numMessage};
+        unsigned char udhs[] = {0x00, 0x03, (uint8_t)concatenatedMessageId,
+                                (uint8_t)numMessages, ++numMessage};
         UserDataHeader udh(std::string((char*)udhs, 5));
         smsTemplate->setUserDataHeader(udh);
       }
